LLM-Flask/requirement.txt
flask
flask-cors
flask_cors
langchain
langchain_community
langchain_openai
langchain_core
chromadb
pymongo
python-dotenv